think I got the author in
I just simply had it pass the String user from login into createpost and
used that.

It might just be because I'm tired, but I thought I remembered the
program trying to log you in before take you to the menu with the option
to login. If it did do that, it doesn't do that anymore and I'm not sure
why. If it didn't do that, then yay it works.