subscription stuff
changed option 3 on login menu from trending to 'choose a user to see
the posts of'

Option 3 now asks for a user to be inputted and then sends it to the
findPost function in createPost

changed 'findPost' function in createPost class to accept an argument,
it *should* find all posts authored by the selected user and print them
out, but it does not.

added stuff to CreateAccount, when a new account is created and a
username is selected, it appends it to subscription.txt, which just
keeps a list of all usernames, nothing else.

(not yet implemented) choosing option 3 from login menu should print out
all the names in subscriptions so that a user knows what he may pick
from.